Implement Widget Struct
