+++ src/proof_system/preprocessed_circuit.rs
@@ -51,6 +51,7 @@ impl PreProcessedCircuit {
         transcript.append_commitment(b"q_arith", &self.arithmetic.q_arith.commitment);
         transcript.append_commitment(b"q_range", &self.range.q_range.commitment);
         transcript.append_commitment(b"q_logic", &self.logic.q_logic.commitment);
+        transcript.append_commitment(b"q_ecc", &self.ecc.q_ecc.commitment);
 
         transcript.append_commitment(b"left_sigma", &self.permutation.left_sigma.commitment);
         transcript.append_commitment(b"right_sigma", &self.permutation.right_sigma.commitment);
@@ -168,6 +169,7 @@ impl<'de> Deserialize<'de> for PreProcessedCircuit {
                     logic: logic_widg,
                     range: range_widg,
                     permutation: perm_widg,
+                    ecc: ecc_widg,
                     v_h_coset_4n,
                 })
             }
